allow running as root, add permissions checks
In order to have sudo not barf, the file system needs to be mounted with
"-oallow_other", so hardcode that option.
Since "allow_other" is bad in conjunction with "running as root", add
permission checks to every function checking if the caller is really
root, and disallow everything if not (it should be sufficient to do this
only for the "open"-type functions, because everything else will not be
called if open failed, but better be safe than sorry).